Add ssh public key to Fuel master node
This commit adds an ability to use ssh public key authentication for
the Fuel master node by uploading your ssh public key to it. The path
to the public key file can be provided in the 'config.sh' file.

1. add pulibc_key parameter to config.sh
2. create add_authorized_key function at functions/product.sh which
will echo publickey to .ssh/authorized_keys@fuel-master
3. call add_authorized_key at the end of
actions/master-node-create-and-install.sh

Closes-Bug: #1579123
Change-Id: I93560126940457518fc4385c0518ab00c0cac01a