Add tool for sending Alt-SysRq to VMs
This tool is a helper for sending Alt+SysRq+<key> combination
to the guest OS in VirtualBox VM, which is used in Linux for
printing different sorts of debug messages to the kernel log,
like list of processes, stack traces of CPUs and so on, or for
performing actions: reboot, sync e.t.c. It's very useful when
all other ways to communicate to the VM become unavailable.

Change-Id: I63864d8e575a73f5efec6e55cc02483a59935307